Renaming the URDF folder to a more appropriate name and moving the markers to a separate subfolder. (#1073)
(cherry picked from commit 556a7430ada5865ee9cdfa84a1b2b0036832b2e4)
