+++ Gems/ROS2RobotImporter/Code/Source/RobotImporter/RobotImporterWidget.h
@@ -17,7 +17,7 @@
 #include "Pages/RobotDescriptionPage.h"
 #include "Pages/XacroParamsPage.h"
 
-#include "URDF/SdfPrefabMaker.h"
+#include "Building/SdfPrefabMaker.h"
 #include <AzCore/Asset/AssetCommon.h>
 #include <AzCore/std/containers/unordered_map.h>
 #include <AzCore/std/parallel/thread.h>
