+++ Gems/ROS2RobotImporter/Code/Source/SdfAssetBuilder/SdfAssetBuilder.cpp
@@ -24,8 +24,8 @@
 
 #include <RobotImporter/Assets/AssetLookup.h>
 #include <RobotImporter/Assets/AssetPathResolver.h>
+#include <RobotImporter/Building/SdfPrefabMaker.h>
 #include <RobotImporter/Parsing/SdfParser.h>
-#include <RobotImporter/URDF/SdfPrefabMaker.h>
 #include <RobotImporter/Utils/ErrorUtils.h>
 #include <SdfAssetBuilder/SdfAssetBuilderSettings.h>
